Convert SWICS methods to new downloader

+++ heliopy/data/util.py
@@ -49,6 +49,10 @@ class Downloader:
       that interval.
     - :meth:`Downloader.load_local_file()`: given an interval, load the local
       file and return a :class:`pandas.DataFrame` object containing the data.
+
+    Attributes
+    ----------
+    units : dict
     """
     def load(self, starttime, endtime):
         """
@@ -73,8 +77,9 @@ class Downloader:
                 try:
                     dl_path = self.download(interval)
                     local_path.parent.mkdir(parents=True, exist_ok=True)
-                    shutil.copy(dl_path, local_path)
-                    os.remove(dl_path)
+                    if dl_path != local_path:
+                        shutil.copy(dl_path, local_path)
+                        os.remove(dl_path)
                 except NoDataError:
                     continue
 
@@ -89,9 +94,11 @@ class Downloader:
         # Attach units
         if local_path.suffix == '.cdf':
             cdf = _load_local(local_path)
-            units = cdf_units(cdf, manual_units=self.units)
+            self.units = cdf_units(cdf, manual_units=self.units)
+        if not hasattr(self, 'warn_missing_units'):
+            self.warn_missing_units = True
         return units_attach(
-            data, units, warn_missing_units=self.warn_missing_units)
+            data, self.units, warn_missing_units=self.warn_missing_units)
 
     def local_path(self, interval):
         local_path = self.local_dir(interval) / self.fname(interval)
@@ -139,7 +146,8 @@ class Downloader:
 
     def local_dir(self, interval):
         """
-        Local directory for a given interval.
+        Local directory for a given interval. This is relative to the base
+        HelioPy data directory.
 
         Parameters
         ----------
@@ -910,13 +918,13 @@ def _download_remote_unknown_version(
 
 
 def _download_remote(remote_url, filename, local_dir):
-    local_dir = path.Path(local_dir)
+    dl_path = path.Path(local_dir) / filename
     remote_url = _fix_url(remote_url)
     remote_url = remote_url + '/' + filename
-    print('Downloading', remote_url)
-    urlreq.urlretrieve(remote_url,
-                       filename=str(local_dir / filename),
-                       reporthook=_reporthook)
+    print(f'Downloading {remote_url} to {dl_path}')
+    fname, _ = urlreq.urlretrieve(remote_url,
+                                  filename=str(dl_path),
+                                  reporthook=_reporthook)
     print('\n')
